Refactor file transfer logic to dedicated module
