+++ crates/core/tedge_agent/src/file_transfer_server/mod.rs
@@ -1,4 +1,6 @@
+//TODO Rename this module to tedge_server
 pub mod actor;
 pub mod error;
+mod file_transfer;
 pub mod http_rest;
 mod request_files;
